# go-swagger/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "kampus merdeka API Docs",
        "contact": {}
    },
    "paths": {
        "/orders": {
            "get": {
                "description": "creater new order with input on payload",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "get all order",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.Order"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "creater new order with input on payload",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Create an order",
                "parameters": [
                    {
                        "description": "create order",
                        "name": "order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.Order"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.Order"
                        }
                    }
                }
            }
        },
        "/orders/{orderID}": {
            "get": {
                "description": "creater new order with input on payload",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "get detail order",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id order",
                        "name": "orderId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.Order"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.Item": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "item_id example:": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                }
            }
        },
        "main.Order": {
            "type": "object",
            "properties": {
                "customer_name": {
                    "type": "string",
                    "example": "Ali"
                },
                "items": {
                    "$ref": "#/definitions/main.Item"
                },
                "order_id": {
                    "type": "string",
                    "example": "1"
                },
                "ordered_at": {
                    "type": "string"
                }
            }
        }
    }
}